Add Square function to Chess
Chess is an abstraction layer of board and it has also implemented the chess rules. So, I didn't think that this layer needs to provide any information of the Board. But, I realize that for certain rules you should know, for example, a piece of board. So, I add this function that it also has a little abstraction because it doesn't know about Coordinates.